requirements.txt
python-docx>=1.1.0
pywin32>=306
python-dotenv>=1.0.0
